+++ GhostRunner.Server.DAL.Interface/ITaskDataAccess.cs
@@ -12,9 +12,9 @@ namespace GhostRunner.Server.DAL.Interface
 
         Task GetNextUnprocessed();
 
-        Boolean SetTaskProcessing(int taskId);
+        Boolean SetTaskProcessing(long taskId);
 
-        Boolean SetTaskComplete(int taskId, Status status);
+        Boolean SetTaskComplete(long taskId, Status status);
 
         Task Insert(Task initializationTask);
     }
